Removed class RoomsServiceUtils because of its contents too specific for RoomServiceVerticle than now has all its methods inside
